+++ sdk/common/utils.d.ts
@@ -7,7 +7,7 @@ export function findBestMatchingLink(allLinks?: any[], pathname?: string): {
     value: string;
     params: {};
 };
-export function convertStringToBase64(string: any): any;
+export function convertStringToBase64(string: any): string;
 export function isBrowser(): boolean;
 export function isNode(): boolean;
 export namespace NAV_TYPE {