+++ sdk/partner/Webhook/WebhookPartnerValidator.d.ts
@@ -5,9 +5,13 @@ declare class WebhookValidator {
     static getDeliveryDetailInsights(): any;
     static fetchDeliveryTs(): any;
     static fetchReportFilters(): any;
+    static downloadDeliveryReport(): any;
     static cancelReportDownload(): any;
     static getHistoricalReports(): any;
     static getInvalidEventList(): any;
     static fetchSubscribers(): any;
     static updateSubscriber(): any;
+    static validateFilterConfiguration(): any;
+    static validateReducerConfiguration(): any;
+    static saveFilterReducerConfiguration(): any;
 }